backend/pyproject.toml
[project]
name = "geo-agent-backend"
version = "0.1.0"
description = "GEO Agent - Generative Engine Optimization backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "tavily-python>=0.5.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiohttp>=3.10.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
